Patch to work on nixos
I couldn't get this to run on nixos because it was downloading a linux
binary that would not link correctly against nixos's library layout.
There is a patched version available as a flake in 'jcdickinson/codeium.nvim'
repository. If a binary is installed as 'codeium-lsp' on your PATH,
this patched version will use that instead of the auto-downloaded one.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>